Improvements to compat to deal with ordered dict properly
